+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 10.9
 ---------------
 - Fix do not add new Enrollment Record if existing has no Dropped date in Student.php, Enrollment.inc.php & SaveEnrollment.fnc.php
 - Enrollment Start: No N/A option for first entry in StudentsUsersInfo.fnc.php
+- Hide End Date input for Inactive Students (no Attendance Start Date) in StudentsUsersInfo.fnc.php
 - SQL ORDER BY DEFAULT_CALENDAR IS NULL,DEFAULT_CALENDAR ASC (nulls first) for consistency between PostgreSQL & MySQL in Courses.fnc.php, Calendar.php & Enrollment.inc.php
 - Simplify & harmonize code, use GetAllMP() in GPARankList.php, InputFinalGrades.php & TeacherCompletion.php
 - Exclude already associated parents/students from Search() in AddUsers.php & AddStudents.php